feat: move `output.BomRefDiscriminator` to `contrib.bom.utils.BomRefDiscriminator`
