+++ cyclonedx/output/__init__.py
@@ -22,17 +22,22 @@ and according to different versions of the CycloneDX schema standard.
 """
 
 import os
+import sys
 from abc import ABC, abstractmethod
-from collections.abc import Iterable, Mapping
-from itertools import chain
-from random import random
+from collections.abc import Mapping
 from typing import TYPE_CHECKING, Any, Literal, Optional, Union, overload
+
+if sys.version_info >= (3, 13):
+    from warnings import deprecated
+else:
+    from typing_extensions import deprecated
 
 from ..schema import OutputFormat, SchemaVersion
 
+from ..contrib.bom.utils import BomRefDiscriminator as _BomRefDiscriminator
+
 if TYPE_CHECKING:  # pragma: no cover
     from ..model.bom import Bom
-    from ..model.bom_ref import BomRef
     from .json import Json as JsonOutputter
     from .xml import Xml as XmlOutputter
 
@@ -138,40 +143,18 @@ def make_outputter(bom: 'Bom', output_format: OutputFormat, schema_version: Sche
         raise ValueError(f'Unknown {output_format.name}/schema_version: {schema_version!r}')
     return klass(bom)
 
+# region deprecated re-export
 
-class BomRefDiscriminator:
+@deprecated('Deprecated re-export location - see docstring of "BomRefDiscriminator" for details.')
+class BomRefDiscriminator(_BomRefDiscriminator):
+    """Deprecated — Alias of :class:`cyclonedx.contrib.license.factories.LicenseFactory`.
 
-    def __init__(self, bomrefs: Iterable['BomRef'], prefix: str = 'BomRef') -> None:
-        # do not use dict/set here, different BomRefs with same value have same hash and would shadow each other
-        self._bomrefs = tuple((bomref, bomref.value) for bomref in bomrefs)
-        self._prefix = prefix
+    .. deprecated:: next
+        This re-export location is deprecated.
+        Use ``from cyclonedx.contrib.bom.utils import BomRefDiscriminator`` instead.
+        The exported symbol itself is NOT deprecated — only this import path.
+    """
+    pass
 
-    def __enter__(self) -> None:
-        self.discriminate()
 
-    def __exit__(self, exc_type: Any, exc_val: Any, exc_tb: Any) -> None:
-        self.reset()
-
-    def discriminate(self) -> None:
-        known_values = []
-        for bomref, _ in self._bomrefs:
-            value = bomref.value
-            if value is None or value in known_values:
-                value = self._make_unique()
-                bomref.value = value
-            known_values.append(value)
-
-    def reset(self) -> None:
-        for bomref, original_value in self._bomrefs:
-            bomref.value = original_value
-
-    def _make_unique(self) -> str:
-        return f'{self._prefix}{str(random())[1:]}{str(random())[1:]}'  # nosec B311
-
-    @classmethod
-    def from_bom(cls, bom: 'Bom', prefix: str = 'BomRef') -> 'BomRefDiscriminator':
-        return cls(chain(
-            map(lambda c: c.bom_ref, bom._get_all_components()),
-            map(lambda s: s.bom_ref, bom.services),
-            map(lambda v: v.bom_ref, bom.vulnerabilities)
-        ), prefix)
+# endregion deprecated re-export
